Add some failing test cases
