+++ attr_test.go
@@ -281,100 +281,3 @@ func TestBoundary(t *testing.T) {
 		})
 	}
 }
-
-func TestConvexHull(t *testing.T) {
-	for i, tt := range []struct {
-		input  string
-		output string
-	}{
-		{
-			input:  "POINT EMPTY",
-			output: "POINT EMPTY",
-		},
-
-		// Fails
-		//{
-		//input:  "POINT(1 2)",
-		//output: "POINT(1 2)",
-		//},
-
-		{
-			input:  "LINESTRING EMPTY",
-			output: "LINESTRING EMPTY",
-		},
-
-		// Fails
-		//{
-		//input:  "LINESTRING(1 2,3 4)",
-		//output: "LINESTRING(1 2,3 4)",
-		//},
-
-		{
-			input:  "LINESTRING(0 0,1 1,1 0,0 1)",
-			output: "POLYGON((0 0,1 0,1 1,0 1,0 0))",
-		},
-
-		{
-			input:  "POLYGON((0 0,0 1,1 0,0 0))",
-			output: "POLYGON((0 0,1 0,0 1,0 0))",
-		},
-		{
-			input:  "POLYGON((0 0,2 0,1 1,2 2,0 2,0 0))",
-			output: "POLYGON((0 0,2 0,2 2,0 2,0 0))",
-		},
-		{
-			input:  "POLYGON EMPTY",
-			output: "POLYGON EMPTY",
-		},
-		{
-			input:  "MULTIPOINT(0 0,0 1,1 0)",
-			output: "POLYGON((0 0,1 0,0 1,0 0))",
-		},
-		{
-			input:  "MULTIPOINT EMPTY",
-			output: "MULTIPOINT EMPTY",
-		},
-		{
-			input:  "MULTILINESTRING EMPTY",
-			output: "MULTILINESTRING EMPTY",
-		},
-
-		// Doesn't pass.
-		//{
-		//input:  "MULTILINESTRING((0 1,2 3))",
-		//output: "LINESTRING(0 1,2 3)",
-		//},
-
-		{
-			input:  "MULTIPOLYGON EMPTY",
-			output: "MULTIPOLYGON EMPTY",
-		},
-		{
-			input:  "MULTIPOLYGON(((0 0,1 0,0 1,0 0)))",
-			output: "POLYGON((0 0,1 0,0 1,0 0))",
-		},
-		{
-			input:  "GEOMETRYCOLLECTION EMPTY",
-			output: "GEOMETRYCOLLECTION EMPTY",
-		},
-
-		// Fails
-		//{
-		//input:  "GEOMETRYCOLLECTION(POINT(1 2))",
-		//output: "POINT(1 2)",
-		//},
-		//{
-		//input:  "GEOMETRYCOLLECTION(GEOMETRYCOLLECTION(POINT(1 2)),POINT(2 1))",
-		//output: "LINESTRING(1 2,2 1)",
-		//},
-		//{
-		//input:  "GEOMETRYCOLLECTION(LINESTRING(1 2,3 4))",
-		//output: "LINESTRING(1 2,3 4)",
-		//},
-	} {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			got := geomFromWKT(t, tt.input).ConvexHull()
-			expectDeepEqual(t, got, geomFromWKT(t, tt.output))
-		})
-	}
-}
